requirements.txt
contourpy
Cython
matplotlib
numpy
pybaselines
pyicoshift
scipy
vam.whittaker

